This is a major overhaul and a replacement of pretty much the whole codebase.
This version 2.0 is now fully object oriented and fully compatible to
GNU sed v4.8 with all it's bells and whissles.

The only things not implemented are character classes (no support in
Python regex) and the command e (might still come in a later update).

The case-modifying escapes \L, \l, \U, \u and \E in the replacement
string are now supported as well.

The GNU sed v4.8 documentation can be used as a reference for pythonsed.

All testcase - except the two for character classes and command e -
are passing now.

Testcases for the new commands v,R and W have been added as well as
a test script (test_script_inplace.py) to test the in-place editiing.

I also added the possibility to use the much more powerful Python
syntax for the regex in addresses and s command by specifying the
-p or --python-syntax option.

Note the different handling of empty matches in those cases, depending
on the Python version and if -p is used or not...